feat(folio): adding update occurence cards in main folio endpoint
