+++ start/routes.ts
@@ -6,6 +6,7 @@ const ApiInfoController = () => import('#controllers/system/api_info_controller'
 const CardsController = () => import('#controllers/product/cards_controller')
 const ScanController = () => import('#controllers/product/scan_controller')
 const FoliosController = () => import('#controllers/product/folios_controller')
+const CardFoliosController = () => import('#controllers/product/card_folios_controller')
 
 router.get('/', [RootController])
 router.get('/health', [HealthCheckController])
@@ -25,7 +26,10 @@ router
         router.post('/scan/analyze', [ScanController, 'analyze'])
         router.post('/folios/init', [FoliosController, 'init']).use(middleware.auth())
         router.get('/folios/cards', [FoliosController, 'cards']).use(middleware.auth())
-        router.post('/folios/collect', [FoliosController, 'collect']).use(middleware.auth())
+        router.post('/folios/collect', [CardFoliosController, 'collect']).use(middleware.auth())
+        router
+          .patch('/folios/occurence', [CardFoliosController, 'occurrence'])
+          .use(middleware.auth())
         router.get('/folios/statistics', [FoliosController, 'statistics']).use(middleware.auth())
       })
       .prefix('v1')
